pyproject.toml
[project]
name = "Reminder-App"
version = "0.1.0"
description = "A simple reminder system for Windows using customtkinter."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=11.3.0",
    "pyinstaller>=6.15.0",
    "pystray>=0.19.5",
    "schedule>=1.2.2",
    "winotify>=1.1.0",
]
